+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- refactor(footer): replace button elements with FooterButton and FooterIcon components. (`9fbddbf`)
 - refactor(footer): streamline footer panel components and remove unused utilities. (`8b68b35`)
 - feat(form): introduce FormGroup component for consistent form layout. (`9708ab0`)
 - feat(modal): enhance modal components with new structure and features. (`c4509d7`)